\name{InfluenceFunctions-package}
\alias{InfluenceFunctions-package}
\alias{InfluenceFunctions}
\docType{package}
\title{
\packageTitle{InfluenceFunctions}
}
\description{
\packageDescription{InfluenceFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{InfluenceFunctions}
\packageIndices{InfluenceFunctions}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{InfluenceFunctions}

Maintainer: \packageMaintainer{InfluenceFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
